+++ packaging/homebrew/agentgate.rb
@@ -2,7 +2,7 @@
 cask "agentgate" do
   arch arm: "aarch64", intel: "x64"
 
-  version "2.0.0"
+  version "2.0.1"
   sha256 arm:   "3146f681cfe80ed6229b545c78e2fb7c212e21a24aebf67488bae10684bc5abf",
          intel: "6f015060bc8ef60c49e782fb9e1250c915698bf2b5259f27dae481ed8c4bf3ea"